将StateMachine trait和enum StateCallback移动到state.rs
1
