+++ src/utils/ui/mode/mod.rs
@@ -1,11 +1,4 @@
-use mode::ModeType;
-
 pub mod common;
 pub mod mode;
 pub mod normal;
-
-pub enum StateCallback {
-    None,
-    Reset,
-    Exit(ModeType),
-}
+pub mod state;
